JMAP: moved _header_as_foo() functions from jmap_mail.c into jmap_util.c as jmap_header_as_foo()
Also, jmap_header_as_date() has been changes to emit a 'Date' type rather than a 'UTCDate' type
